# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "movie-reviews-api",
        "description": "API documentation movie-reviews-api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reviews/{review_id}/replies": {
            "get": {
                "tags": [
                    "reply"
                ],
                "summary": "get replies of specified review.",
                "operationId": "d2c28eab55736f1695ab7b226024dbd8",
                "parameters": [
                    {
                        "name": "review_id",
                        "in": "path",
                        "description": "id of review",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    },
                    {
                        "name": "search_term",
                        "in": "query",
                        "description": "search_term",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "username",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "likes_count",
                        "in": "query",
                        "description": "filter likes_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "dislikes_count",
                        "in": "query",
                        "description": "filter dislikes_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "replies_count",
                        "in": "query",
                        "description": "filter replies_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "number of items in a page",
                        "schema": {
                            "format": "int64",
                            "default": 10
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort by",
                        "schema": {
                            "format": "string",
                            "default": "",
                            "enum": [
                                "newest",
                                "oldest",
                                "most-likes",
                                "least-likes",
                                "most-dislikes",
                                "least-dislikes",
                                "most-replies",
                                "least-replies"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/replies/{reply_id}/replies": {
            "get": {
                "tags": [
                    "reply"
                ],
                "summary": "get replies of specified reply.",
                "operationId": "a3b1d3a700976b57cbb9fce32a40fba8",
                "parameters": [
                    {
                        "name": "reply_id",
                        "in": "path",
                        "description": "id of reply",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    },
                    {
                        "name": "search_term",
                        "in": "query",
                        "description": "search_term",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "username",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "likes_count",
                        "in": "query",
                        "description": "filter likes_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "dislikes_count",
                        "in": "query",
                        "description": "filter dislikes_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "replies_count",
                        "in": "query",
                        "description": "filter replies_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "number of items in a page",
                        "schema": {
                            "format": "int64",
                            "default": 10
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort by",
                        "schema": {
                            "format": "string",
                            "default": "",
                            "enum": [
                                "newest",
                                "oldest",
                                "most-likes",
                                "least-likes",
                                "most-dislikes",
                                "least-dislikes",
                                "most-replies",
                                "least-replies"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/replies/{id}": {
            "get": {
                "tags": [
                    "reply"
                ],
                "summary": "get specified reply.",
                "operationId": "727571e68be10d3a0af0ea4968e2cbdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of reply",
                        "required": true,
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "reply not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "reply found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/replies": {
            "post": {
                "tags": [
                    "reply"
                ],
                "summary": "create a reply.",
                "operationId": "f988880d3d29547742972a3783101a60",
                "requestBody": {
                    "description": "request body",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "description": "reply text",
                                        "default": ""
                                    },
                                    "review_id": {
                                        "description": "",
                                        "default": ""
                                    },
                                    "reply_id": {
                                        "description": "",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "reply created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "media is already reviewd ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/replies/{reply_id}": {
            "put": {
                "tags": [
                    "reply"
                ],
                "summary": "update specified review.",
                "operationId": "9762d79e42d3af7ad8100a3c9e843e9f",
                "parameters": [
                    {
                        "name": "reply_id",
                        "in": "path",
                        "description": "id of reply",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "request body",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "description": "new reply text",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "reply updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "reply not found or not created by user",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "reply"
                ],
                "summary": "delete specified reply.",
                "operationId": "1bf64db999118ea1802bb9b521f5426f",
                "parameters": [
                    {
                        "name": "reply_id",
                        "in": "path",
                        "description": "id of reply",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "reply deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "reply not found or not created by user",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}